+++ src/components/base/Base.js
@@ -8,6 +8,7 @@ import Formio from '../../Formio';
 import Validator from '../Validator';
 import Widgets from '../../widgets';
 import Component from '../../Component';
+import dragula from 'dragula';
 const CKEDITOR = 'https://cdn.staticaly.com/gh/formio/ckeditor5-build-classic/master/build/ckeditor.js';
 
 /**
@@ -891,11 +892,17 @@ export default class BaseComponent extends Component {
     if (!this.tbody) {
       return;
     }
+    const hasDraggableRows = this.hasDraggableRows;
     this.inputs = [];
     this.tbody.innerHTML = '';
     values = values || this.dataValue;
     _.each(values, (value, index) => {
       const tr = this.ce('tr');
+      if (hasDraggableRows) {
+        tr.appendChild(this.ce('td', {
+          class: 'formio-drag-column'
+        }, this.dragButton()));
+      }
       const td = this.ce('td');
       this.buildInput(td, value, index);
       tr.appendChild(td);
@@ -906,13 +913,18 @@ export default class BaseComponent extends Component {
         tr.appendChild(tdAdd);
       }
 
+      if (hasDraggableRows) {
+        tr.dragInfo = {
+          index: index
+        };
+      }
       this.tbody.appendChild(tr);
     });
 
     if (!this.shouldDisable) {
       const tr = this.ce('tr');
       const td = this.ce('td', {
-        colspan: '2'
+        colspan: hasDraggableRows ? '3' : '2'
       });
       td.appendChild(this.addButton());
       tr.appendChild(td);
@@ -922,6 +934,48 @@ export default class BaseComponent extends Component {
     if (this.shouldDisable) {
       this.disabled = true;
     }
+
+    if (hasDraggableRows) {
+      this.addDraggable([this.tbody]);
+    }
+  }
+
+  get hasDraggableRows() {
+    return this.component.hasDraggableRows && !this.options.readOnly;
+  }
+
+  addDraggable(containers) {
+    this.dragula = dragula(containers, this.getRowDragulaOptions()).on('drop', this.onRowDrop.bind(this));
+  }
+
+  getRowDragulaOptions() {
+    return {
+      moves: function(draggedElement, oldParent, clickedElement) {
+        //allow dragging only on drag button (not the whole row)
+        return clickedElement.classList.contains('formio-drag-button');
+      }
+    };
+  }
+
+  onRowDrop(droppedElement, newParent, oldParent, nextSibling) {
+    //move them in data value as well
+    if (!droppedElement.dragInfo || (nextSibling && !nextSibling.dragInfo)) {
+      console.warn('There is no Drag Info available for either dragged or sibling element');
+      return;
+    }
+    const oldPosition = droppedElement.dragInfo.index;
+    //should drop at next sibling position; no next sibling means drop to last position
+    const newPosition = nextSibling ? nextSibling.dragInfo.index : this.dataValue.length;
+    const movedBelow = newPosition > oldPosition;
+    const dataValue = _.cloneDeep(this.dataValue);
+    const draggedRowData = dataValue[oldPosition];
+
+    //insert element at new position
+    dataValue.splice(newPosition, 0, draggedRowData);
+    //remove element from old position (if was moved above, after insertion it's at +1 index)
+    dataValue.splice(movedBelow ? oldPosition : oldPosition + 1, 1);
+    //need to re-build rows to re-calculate indexes and other indexed fields for component instance (like rows for ex.)
+    this.setValue(dataValue);
   }
 
   buildInput(container, value) {
@@ -1007,6 +1061,12 @@ export default class BaseComponent extends Component {
     });
     removeButton.appendChild(removeIcon);
     return removeButton;
+  }
+
+  dragButton() {
+    return this.ce('button', {
+      class: `formio-drag-button btn btn-default btn-small ${this.iconClass('resize-vertical')}`
+    });
   }
 
   labelOnTheLeft(position) {