+++ src/components/columns/editForm/Columns.edit.display.js
@@ -7,6 +7,7 @@ export default [
     label: 'Column Properties',
     addAnother: 'Add Column',
     tooltip: 'The width, offset, push, and pull settings for each column.',
+    hasDraggableRows: true,
     components: [
       {
         type: 'number',