+++ src/components/edittable/EditTable.form.js
@@ -37,6 +37,7 @@ export default function(...extend) {
         tableView: true,
         alwaysEnabled: false,
         type: 'datagrid',
+        hasDraggableRows: true,
         input: true,
         key: 'rowGroups',
         components: [