+++ src/components/edittable/editForm/EditTable.edit.display.js
@@ -5,6 +5,7 @@ export default [
     input: true,
     label: 'Columns',
     weight: 100,
+    hasDraggableRows: true,
     components: [
       {
         key: 'label',