+++ src/components/select/editForm/Select.edit.data.js
@@ -88,6 +88,7 @@ export default [
     key: 'data.values',
     tooltip: 'Values to use as the data source. Labels are shown in the select field. Values are the corresponding values saved with the submission.',
     weight: 10,
+    hasDraggableRows: true,
     components: [
       {
         label: 'Label',